Add files via upload
This is my 3rd year group project of Engineering Product Design. Under the project, we have designed an electric plug socket for the domestic users which can be controlled by a time control setting. This device can be assembled with the home installations system & can be used to reduce unnecessary energy consumption & time wastage.